feat: add the EP1 files
In this EP, the user receives as input a matrix filled with "water" (number zero) and "garbage" (numbers greater than zero). Next, we need to count the "islands" (garbage joined by one horizontal/vertical call), the amount of garbage each islands has, and the a list of island that perfectly fill a "container" with arbitrary capacity.
The program generate_matrix.c is a tool to generate random matrix to use as test cases in the main program.